requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
boto3==1.36.17
botocore==1.36.17
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dateparser==1.2.1
exceptiongroup==1.2.2
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
numpy==2.2.2
pandas==2.2.3
propcache==0.2.1
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
